chore: move image component
