+++ apps/web/src/components/Home/Timeline/EventType/ProfileCircles.tsx
@@ -1,7 +1,7 @@
-import { Image } from '@components/UI/Image';
 import type { Profile } from 'lens';
 import Link from 'next/link';
 import type { FC, ReactNode } from 'react';
+import { Image } from 'ui';
 import formatHandle from 'utils/formatHandle';
 import getAvatar from 'utils/getAvatar';
