+++ apps/web/src/components/Messages/MessagesList.tsx
@@ -1,5 +1,4 @@
 import Markup from '@components/Shared/Markup';
-import { Image } from '@components/UI/Image';
 import { EmojiSadIcon } from '@heroicons/react/outline';
 import { formatTime } from '@lib/formatTime';
 import { Trans } from '@lingui/macro';
@@ -10,7 +9,7 @@ import type { Profile } from 'lens';
 import type { FC, ReactNode } from 'react';
 import { memo } from 'react';
 import { useInView } from 'react-cool-inview';
-import { Card } from 'ui';
+import { Card, Image } from 'ui';
 import formatHandle from 'utils/formatHandle';
 import getAvatar from 'utils/getAvatar';
