+++ apps/web/src/components/Notification/WalletProfile.tsx
@@ -1,7 +1,7 @@
-import { Image } from '@components/UI/Image';
 import { POLYGONSCAN_URL } from 'data/constants';
 import type { Wallet } from 'lens';
 import type { FC } from 'react';
+import { Image } from 'ui';
 import formatAddress from 'utils/formatAddress';
 import getStampFyiURL from 'utils/getStampFyiURL';
 import imageProxy from 'utils/imageProxy';
