+++ apps/web/src/components/Shared/IFramely/Embed.tsx
@@ -1,4 +1,3 @@
-import { Image } from '@components/UI/Image';
 import { Mixpanel } from '@lib/mixpanel';
 import { stopEventPropagation } from '@lib/stopEventPropagation';
 import { ATTACHMENT } from 'data/constants';
@@ -6,7 +5,7 @@ import Link from 'next/link';
 import type { FC } from 'react';
 import { PUBLICATION } from 'src/tracking';
 import type { OG } from 'src/types';
-import { Card } from 'ui';
+import { Card, Image } from 'ui';
 import imageProxy from 'utils/imageProxy';
 
 interface EmbedProps {
