+++ apps/web/src/components/Shared/Navbar/MobileDrawerMenu.tsx
@@ -1,4 +1,3 @@
-import { Image } from '@components/UI/Image';
 import { XIcon } from '@heroicons/react/outline';
 import { Trans } from '@lingui/macro';
 import type { Profile } from 'lens';
@@ -6,6 +5,7 @@ import Link from 'next/link';
 import type { FC } from 'react';
 import { useAppStore } from 'src/store/app';
 import { useGlobalModalStateStore } from 'src/store/modals';
+import { Image } from 'ui';
 import formatHandle from 'utils/formatHandle';
 import getAvatar from 'utils/getAvatar';
 import isGardener from 'utils/isGardener';
