+++ apps/web/src/components/Shared/Navbar/SignedUser.tsx
@@ -1,4 +1,3 @@
-import { Image } from '@components/UI/Image';
 import { Menu } from '@headlessui/react';
 import { Trans } from '@lingui/macro';
 import clsx from 'clsx';
@@ -6,6 +5,7 @@ import type { Profile } from 'lens';
 import type { FC } from 'react';
 import { useAppStore } from 'src/store/app';
 import { useGlobalModalStateStore } from 'src/store/modals';
+import { Image } from 'ui';
 import formatHandle from 'utils/formatHandle';
 import getAvatar from 'utils/getAvatar';
 import isGardener from 'utils/isGardener';
