+++ packages/ui/index.ts
@@ -7,3 +7,4 @@ export { ErrorMessage } from './ErrorMessage';
 export { FieldError, Form, useZodForm } from './Form';
 export { GridItemEight, GridItemFour, GridLayout } from './GridLayout';
 export { default as HelpTooltip } from './HelpTooltip';
+export { Image } from './Image';
